requirements.txt
yfinance>=0.2.36
pandas>=2.0
numpy>=1.24
reportlab>=4.0
matplotlib>=3.7
resend>=2.0
python-dotenv>=1.0
